activity feed backend

+++ app/Http/Controllers/ThreadsController.php
@@ -110,7 +110,14 @@ class ThreadsController extends Controller
      */
     public function destroy(Channel $channel,Thread $thread)
     {
+        $this->authorize('update',$thread);
+
+
         $thread->delete();
+
+        if (request()->wantsJson()){
+            return response([],204);
+        }
 
         return redirect('/threads');
     }
